+++ tasks.py
@@ -504,6 +504,7 @@ def regenerate_package_mode(c, debug=False, swagger_group=None):
         'test/azure/legacy/specification/packagemodemgmtplane/README.md',
         'test/azure/legacy/specification/packagemodecustomize/README.md',
         'test/azure/legacy/specification/packagemodedataplane/README.md',
+        'test/azure/legacy/specification/packagemodebatch/README.md',
     ]
     vanilla_packages = [
         'test/vanilla/legacy/specification/packagemodedataplane/README.md',